+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
+* Added support for IndicatorTypes directory in packs (for `reputation` files, instead of Misc).
+* Fixed a bug in the **validate** command where indicatorField always field.
+* Fixed a bug in the **validate** command where layouts version was not checked correctly.
+
+#### 1.0.3
 * Fixed a bug in the **validate** command where indicatorField always field.
 * Fixed a bug in the **validate** command where layouts version was not checked correctly.
 * Fixed a bug in the **format** command where comments were being removed from YAML files.